+++ packages/client/src/pages/Forum/Forum.module.css
@@ -10,8 +10,3 @@
   top: 45px;
   z-index: 1;
 }
-
-.button.button {
-  margin: 5px;
-  padding: 5px;
-}
