+++ packages/client/src/pages/Forum/Forum.module.scss
@@ -13,8 +13,3 @@
   top: 45px;
   z-index: 1;
 }
-
-.button.button {
-  margin: 5px;
-  padding: 5px;
-}
