+++ packages/client/src/pages/Forum/components/table/ForumsTable.module.css
@@ -1,0 +1,4 @@
+.button.button {
+  margin: 5px;
+  padding: 5px;
+}
